productivity/requirements.txt
ultralytics>=8.2.0
opencv-python==4.12.0.88
numpy==2.2.2
pyyaml>=6.0.1
shapely>=2.0.2
onnxruntime>=1.18.0
supervision>=0.22.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0
